pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etsy-niche-scout"
version = "0.1.0"
description = "Local-first Etsy niche research tool for digital products."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jinja2>=3.1",
  "pandas>=2.2",
  "playwright>=1.52",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "rapidfuzz>=3.9",
  "rich>=13.7",
  "streamlit>=1.44",
  "tenacity>=8.3",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[project.scripts]
niche-scout = "niche_scout.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
